+++ pkg/dds/lib/src/dap/adapters/dart.dart
@@ -8,6 +8,7 @@ import 'dart:io';
 
 import 'package:collection/collection.dart';
 import 'package:dap/dap.dart';
+import 'package:dds_service_extensions/dds_service_extensions.dart';
 import 'package:json_rpc_2/error_code.dart' as json_rpc_errors;
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as path;
@@ -434,12 +435,19 @@ abstract class DartDebugAdapter<TL extends LaunchRequestArguments,
   /// vs 'detached').
   bool isDetaching = false;
 
-  /// Whether isolates that pause in the PauseExit state should be automatically
-  /// resumed after any in-process log events have completed.
+  /// Whether this adapter set the --pause-isolates-on-start flag, specifying
+  /// that isolates should pause on starting.
   ///
   /// Normally this will be true, but it may be set to false if the user
-  /// also manually passes pause-isolates-on-exit.
-  bool resumeIsolatesAfterPauseExit = true;
+  /// also manually passed the --pause-isolates-on-start flag.
+  bool pauseIsolatesOnStartSetByDap = true;
+
+  /// Whether this adapter set the --pause-isolates-on-exit flag, specifying
+  /// that isolates should pause on exiting.
+  ///
+  /// Normally this will be true, but it may be set to false if the user
+  /// also manually passed the --pause-isolates-on-exit flag.
+  bool pauseIsolatesOnExitSetByDap = true;
 
   /// A [Future] that completes when the last queued OutputEvent has been sent.
   ///
@@ -558,10 +566,6 @@ abstract class DartDebugAdapter<TL extends LaunchRequestArguments,
     this.args = args as DartCommonLaunchAttachRequestArguments;
     isAttach = true;
     _subscribeToOutputStreams = true;
-
-    // When attaching to a process, suppress auto-resuming isolates until the
-    // first time the user resumes anything.
-    isolateManager.autoResumeStartingIsolates = false;
 
     // Common setup.
     await _prepareForLaunchOrAttach(null);
@@ -702,6 +706,7 @@ abstract class DartDebugAdapter<TL extends LaunchRequestArguments,
     // Let the subclass do any existing setup once we have a connection.
     await debuggerConnected(vmInfo);
 
+    await _configureIsolateSettings(vmService);
     await _withErrorHandling(
       () => _configureExistingIsolates(vmService, vmInfo),
     );
@@ -756,6 +761,68 @@ abstract class DartDebugAdapter<TL extends LaunchRequestArguments,
     return DapProgressReporter.start(this, id, title, message: message);
   }
 
+  Future<void> _configureIsolateSettings(
+    vm.VmService vmService,
+  ) async {
+    // If this is an attach workflow, check whether pause_isolates_on_start or
+    // pause_isolates_on_exit were already set, and if not set them (note: this
+    // is already done as part of the launch workflow):
+    if (isAttach) {
+      const pauseIsolatesOnStart = 'pause_isolates_on_start';
+      const pauseIsolatesOnExit = 'pause_isolates_on_exit';
+      final flags = (await vmService.getFlagList()).flags ?? <vm.Flag>[];
+      for (final flag in flags) {
+        final flagName = flag.name;
+        final isPauseIsolatesFlag =
+            flagName == pauseIsolatesOnStart || flagName == pauseIsolatesOnExit;
+        if (flagName == null || !isPauseIsolatesFlag) continue;
+
+        if (flag.valueAsString == 'true') {
+          if (flagName == pauseIsolatesOnStart) {
+            pauseIsolatesOnStartSetByDap = false;
+          }
+          if (flagName == pauseIsolatesOnExit) {
+            pauseIsolatesOnExitSetByDap = false;
+          }
+        } else {
+          _setVmFlagTo(vmService, flagName: flagName, valueAsString: 'true');
+        }
+      }
+    }
+
+    try {
+      // Make sure DDS waits for DAP to be ready to resume before forwarding
+      // resume requests to the VM Service:
+      await vmService.requirePermissionToResume(
+        onPauseStart: true,
+        onPauseExit: true,
+      );
+
+      // Specify whether DDS should wait for a user-initiated resume as well as a
+      // DAP-initiated resume:
+      await vmService.requireUserPermissionToResume(
+        onPauseStart: !pauseIsolatesOnStartSetByDap,
+        onPauseExit: !pauseIsolatesOnExitSetByDap,
+      );
+    } catch (e) {
+      // If DDS is not enabled, calling these DDS service extensions will fail.
+      // Therefore catch and log any errors.
+      logger?.call('Failure configuring isolate settings: $e');
+    }
+  }
+
+  Future<void> _setVmFlagTo(
+    vm.VmService vmService, {
+    required String flagName,
+    required String valueAsString,
+  }) async {
+    try {
+      await vmService.setFlag(flagName, valueAsString);
+    } catch (e) {
+      logger?.call('Failed to to set VM flag $flagName to $valueAsString: $e');
+    }
+  }
+
   /// Process any existing isolates that may have been created before the
   /// streams above were set up.
   Future<void> _configureExistingIsolates(
@@ -775,8 +842,7 @@ abstract class DartDebugAdapter<TL extends LaunchRequestArguments,
       final pauseEventKind = isolate.runnable ?? false
           ? vm.EventKind.kIsolateRunnable
           : vm.EventKind.kIsolateStart;
-      final thread =
-          await isolateManager.registerIsolate(isolate, pauseEventKind);
+      await isolateManager.registerIsolate(isolate, pauseEventKind);
 
       // If the Isolate already has a Pause event we can give it to the
       // IsolateManager to handle (if it's PausePostStart it will re-configure
@@ -788,13 +854,7 @@ abstract class DartDebugAdapter<TL extends LaunchRequestArguments,
           isolate.pauseEvent!,
         );
       } else if (isolate.runnable == true) {
-        // If requested, automatically resume. Otherwise send a Stopped event to
-        // inform the client UI the thread is paused.
-        if (isolateManager.autoResumeStartingIsolates) {
-          await isolateManager.resumeIsolate(isolate);
-        } else {
-          isolateManager.sendStoppedOnEntryEvent(thread.threadId);
-        }
+        await isolateManager.readyToResumeIsolate(isolate);
       }
     }));
   }
@@ -2352,11 +2412,9 @@ abstract class DartDebugAdapter<TL extends LaunchRequestArguments,
     // We pause isolates on exit to allow requests for resolving URIs in
     // stderr call stacks, so when we see an isolate pause, wait for any
     // pending logs and then resume it (so it exits).
-    if (resumeIsolatesAfterPauseExit &&
-        eventKind == vm.EventKind.kPauseExit &&
-        isolate != null) {
+    if (eventKind == vm.EventKind.kPauseExit && isolate != null) {
       await _waitForPendingOutputEvents();
-      await isolateManager.resumeIsolate(isolate);
+      await isolateManager.readyToResumeIsolate(isolate);
     }
   }