+++ pkg/dds/test/dap/integration/debug_breakpoints_test.dart
@@ -646,7 +646,8 @@ void main(List<String> args) async {
         client2.start(
           launch: () => client2.attach(
             vmServiceUri: vmServiceUri.toString(),
-            autoResume: false,
+            autoResumeOnEntry: false,
+            autoResumeOnExit: false,
             cwd: dap.testAppDir.path,
           ),
         ),